+++ src/editor/copy-paste.js
@@ -2,7 +2,6 @@ import { EditorState, EditorSelection } from "@codemirror/state"
 import { EditorView } from "@codemirror/view"
 
 import { LANGUAGES } from './languages.js';
-import { setEmacsMarkMode } from "./emacs.js"
 
 
 const languageTokensMatcher = LANGUAGES.map(l => l.token).join("|")
@@ -61,7 +60,7 @@ export const heynoteCopyCut = (editor) => {
         }
         
         // if we're in Emacs mode, we want to exit mark mode in case we're in it
-        setEmacsMarkMode(false)
+        editor.emacsMarkMode = false
 
         // if Editor.deselectOnCopy is set (e.g. we're in Emacs mode), we want to remove the selection after we've copied the text
         if (editor.deselectOnCopy && event.type == "copy") {
@@ -95,7 +94,7 @@ const copyCut = (view, cut, editor) => {
     }
 
     // if we're in Emacs mode, we want to exit mark mode in case we're in it
-    setEmacsMarkMode(false)
+    editor.emacsMarkMode = false
 
     // if Editor.deselectOnCopy is set (e.g. we're in Emacs mode), we want to remove the selection after we've copied the text
     if (editor.deselectOnCopy && !cut) {
@@ -107,6 +106,7 @@ const copyCut = (view, cut, editor) => {
             selection: newSelection,
         }))
     }
+    return true
 }