+++ src/stores/editor-cache.js
@@ -38,6 +38,7 @@ export const useEditorCacheStore = defineStore("editorCache", {
                     fontSize: settingsStore.settings.fontSize,
                     defaultBlockToken: settingsStore.settings.defaultBlockLanguage,
                     defaultBlockAutoDetect: settingsStore.settings.defaultBlockLanguageAutoDetect,
+                    keyBindings: settingsStore.settings.keyBindings,
                 })
             } catch (e) {
                 errorStore.addError("Error! " + e.message)
@@ -122,7 +123,8 @@ export const useEditorCacheStore = defineStore("editorCache", {
                         switch (key) {
                             case "keymap":
                             case "emacsMetaKey":
-                                editor.setKeymap(newSettings.keymap, newSettings.emacsMetaKey)
+                            case "keyBindings":
+                                editor.setKeymap(newSettings.keymap, newSettings.emacsMetaKey, newSettings.keyBindings)
                                 break
                             case "showLineNumberGutter":
                                 editor.setLineNumberGutter(newSettings.showLineNumberGutter)